Build 988 Added a hidden config file Added an enum helper string value Removed null termination from the BinaryReader, because it was broken Added color to the lobby chat Not tested yet, but the power and time, and most of the office actions should be synced now Fixed a lot of bugs Added a time display to log messages

+++ Source/Code/CorePlugin/Alzaitu.Lacewing.Client/Packet/Response/PacketResponseSetName.cs
@@ -14,7 +14,7 @@ namespace Alzaitu.Lacewing.Client.Packet.Response
 
         protected override void ReadResponse(byte[] bytes, long size, int pos, LacewingClient client = null)
         {
-            int length = bytes[pos++];
+            byte length = bytes[pos++];
             size--;
             byte[] data = new byte[length];
             for(int i = 0; i < data.Length; i++)
